requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.5
Django==3.1.7
django-bootstrap-datepicker-plus==3.0.5
django-picklefield==3.0.1
pycparser==2.20
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
